+++ src/components/ThemeToggle.vue
@@ -1,6 +1,6 @@
 <template>
   <div>
-    <h2 class="title-2">{{ $t('theme.header') }}</h2>
+    <h2 class="widget__title">{{ $t('theme.header') }}</h2>
     <div v-for="mode in modes" :key="mode.id" class="ods-form-group">
       <input
         type="radio"
@@ -53,3 +53,8 @@ export default {
   },
 };
 </script>
+<style lang="scss" scoped>
+  .widget__title {
+    padding-bottom: 1rem;
+  }
+</style>